python/requirements.txt
fastapi==0.88.0
uvicorn==0.18.3
python-dotenv==1.1.0
vwo-fme-python-sdk==1.5.0
